# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * KDP Digital Products Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { BookBookType } from './bookBookType';
import type { BookInteriorType } from './bookInteriorType';
import type { BookStatus } from './bookStatus';
import type { BookTrimSize } from './bookTrimSize';

export interface Book {
  id: number;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  niche: string;
  /** @nullable */
  targetAudience?: string | null;
  bookType: BookBookType;
  colorPalette: string;
  trimSize: BookTrimSize;
  dayCount: number;
  interiorType: BookInteriorType;
  authorName: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  templateKey?: string | null;
  includeHabitTracker?: boolean;
  includeWeeklyReview?: boolean;
  status: BookStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  generationProgress: number;
  /** @nullable */
  publishedAt?: Date | null;
  /** @nullable */
  lastPageCount?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}