# lib/api-zod/src/generated/types/bookInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * KDP Digital Products Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { BookInputBookType } from './bookInputBookType';
import type { BookInputInteriorType } from './bookInputInteriorType';
import type { BookInputTrimSize } from './bookInputTrimSize';

export interface BookInput {
  /** @minLength 1 */
  title: string;
  subtitle?: string;
  /** @minLength 1 */
  niche: string;
  targetAudience?: string;
  bookType: BookInputBookType;
  colorPalette: string;
  trimSize: BookInputTrimSize;
  /**
     * @minimum 1
     * @maximum 365
     */
  dayCount: number;
  interiorType: BookInputInteriorType;
  /** @minLength 1 */
  authorName: string;
  category?: string;
  templateKey?: string;
  includeHabitTracker?: boolean;
  includeWeeklyReview?: boolean;
  notes?: string;
}