# lib/api-zod/src/generated/types/bookUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * KDP Digital Products Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { BookUpdateBookType } from './bookUpdateBookType';
import type { BookUpdateInteriorType } from './bookUpdateInteriorType';
import type { BookUpdateStatus } from './bookUpdateStatus';
import type { BookUpdateTrimSize } from './bookUpdateTrimSize';

export interface BookUpdate {
  /** @minLength 1 */
  title?: string;
  subtitle?: string;
  niche?: string;
  targetAudience?: string;
  bookType?: BookUpdateBookType;
  colorPalette?: string;
  trimSize?: BookUpdateTrimSize;
  /**
     * @minimum 1
     * @maximum 365
     */
  dayCount?: number;
  interiorType?: BookUpdateInteriorType;
  authorName?: string;
  category?: string;
  templateKey?: string;
  includeHabitTracker?: boolean;
  includeWeeklyReview?: boolean;
  status?: BookUpdateStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  generationProgress?: number;
  notes?: string;
}